Simple Shooting mechanic
It still need to disappear when it collides or get out of view, will work on it later.
I don't know if there is a problem with the jumping animation with the pause menu, I did the same as what I did for the left and right anim problem. Hope it works